backend/requirements.txt
# local deste arquivo  ~/sge/backend/requirements.txt
Flask==2.3.1
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.10
Flask-RESTPlus==0.13.0